Validation

+++ backend/src/main/java/com/example/digitom/domain/report/Report.java
@@ -26,6 +26,6 @@ public class Report {
     private LocalDate date;
 
     @Column(name = "tom", precision = 5, scale = 2)
-    private BigDecimal tom;
+    private double tom;
 
 }
